[lib] Moved `.image.utils` -> `.utils`
